[完美][master][pep-149] PEP-149 实现PEP平台权限相关接口（角色接口部分） (#151)
* auth relations

* user module design

* 增加逻辑删除

* 增加逻辑删除

* auth's function module include menu & resource

* remove sequenceNumber from resource entity

* remove unused import

* Merge remote-tracking branch 'remotes/王贺/auth' into local-王贺auth

# Conflicts:
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/controller/UsersController.java

* Merge remote-tracking branch 'remotes/王贺/auth' into local-王贺auth

# Conflicts:
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/controller/UsersController.java

* refactor menu controller

* fix

* update to 0.3.4-SNAPSHOT

* merged

* 去掉不用的导入

* 现阶段完成了实现用户的工作

* 添加异常抛出

* 去掉抛出异常

* 完善后台各层代码

* 增加测试

* 增加自定义sql语句

* merged

* 实现功能细节

* 增加测试用例

* build

* 增加了判断是否存在循环继承的代码

* Build Success

* Build Success

* Merge remote-tracking branch 'remotes/公司远程/master' into local-王贺auth-PEP-140

# Conflicts:
#	subprojects/pep-auth/pep-auth-api/src/main/java/com/proper/enterprise/platform/api/auth/service/RoleService.java
#	subprojects/pep-auth/pep-auth-api/src/main/java/com/proper/enterprise/platform/api/auth/service/UserService.java
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/controller/UsersController.java
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/service/impl/AbstractUserServiceImpl.java
#	subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/service/impl/RoleServiceImpl.java
#	subprojects/pep-auth/pep-auth-common/src/test/groovy/com/proper/enterprise/platform/auth/common/controller/UsersControllerTest.groovy

* 增加属性，分离用户的角色、用户组信息到新的接口，等待新接口完成后继续。同时增加了对权限的判断，超级管理员等细节处的判断。

* 完善权限代码结构，增加资源权限判断、角色、用户、用户组，各级别的权限判断。

* 增加测试用例

* 添加@JsonIgnore,阻止转换出现级联循环

* 添加@JsonIgnore,阻止转换出现级联循环

* 添加测试，提高覆盖率

* 添加测试，提高覆盖率

* 添加测试，提高覆盖率

* 修改一些错误

* 修改错误

* 增加测试用例，提高覆盖率

* 根据PR修改代码

* 根据PR修改代码

* 修改权限匹配功能，将检测抽象，增加当前用户检测。增加扩展性，复用性

* 修改错误测试用例

* 修改错误

* 优化代码

* Merge remote-tracking branch

* 更改用户匹配url为真实请求

* 增加对子角色的判断

* 增加测试用例

* 去掉获取子角色方法，改为条件查询

+++ subprojects/pep-auth/pep-auth-common/src/main/java/com/proper/enterprise/platform/auth/common/controller/RolesController.java
@@ -2,6 +2,7 @@ package com.proper.enterprise.platform.auth.common.controller;
 
 import com.proper.enterprise.platform.api.auth.model.*;
 import com.proper.enterprise.platform.api.auth.service.RoleService;
+import com.proper.enterprise.platform.api.auth.service.UserService;
 import com.proper.enterprise.platform.core.controller.BaseController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -17,40 +18,45 @@ public class RolesController extends BaseController {
     @Autowired
     private RoleService roleService;
 
+    @Autowired
+    private UserService userService;
+
     @GetMapping
     public ResponseEntity<Collection<? extends Role>> get(String name, String description, String parentId, String enable) {
-        // TODO 具体实现
+        userService.checkPermission("/auth/roles", RequestMethod.GET);
         return responseOfGet(roleService.getByCondiction(name, description, parentId, enable));
     }
 
     @PostMapping
     public ResponseEntity<Role> create(@RequestBody Map<String, Object> reqMap) {
-        // TODO 具体实现
+        userService.checkPermission("/auth/roles", RequestMethod.POST);
         return responseOfPost(roleService.save(reqMap));
     }
 
     @SuppressWarnings("unchecked")
     @PutMapping
     public ResponseEntity<Collection<? extends Role>> updateEnable(@RequestBody Map<String, Object> reqMap) {
+        userService.checkPermission("/auth/roles", RequestMethod.PUT);
         Collection<String> idList = (Collection<String>)reqMap.get("ids");
         boolean enable = (boolean) reqMap.get("enable");
-        // TODO 具体实现
         return responseOfPut(roleService.updateEanble(idList, enable));
     }
 
     @DeleteMapping
     public ResponseEntity delete(@RequestParam String ids) {
-        // TODO 具体实现
+        userService.checkPermission("/auth/roles", RequestMethod.DELETE);
         return responseOfDelete(roleService.deleteByIds(ids));
     }
 
     @GetMapping(path = "/{roleId}")
     public ResponseEntity<Role> find(@PathVariable String roleId) {
+        userService.checkPermission("/auth/roles/" + roleId, RequestMethod.GET);
         return responseOfGet(roleService.get(roleId));
     }
 
     @PutMapping(path = "/{roleId}")
     public ResponseEntity<Role> update(@PathVariable String roleId, @RequestBody Map<String, Object> reqMap) {
+        userService.checkPermission("/auth/roles/" + roleId, RequestMethod.PUT);
         Role role = roleService.get(roleId);
         if (role != null) {
             reqMap.put("id", roleId);
@@ -61,7 +67,7 @@ public class RolesController extends BaseController {
 
     @GetMapping(path = "/{roleId}/menus")
     public ResponseEntity<Collection<? extends Menu>> getRoleMenus(@PathVariable String roleId) {
-        // TODO 具体业务实现
+        userService.checkPermission("/auth/roles/" + roleId + "/menus", RequestMethod.GET);
         return responseOfGet(roleService.getRoleMenus(roleId));
     }
 
@@ -69,14 +75,14 @@ public class RolesController extends BaseController {
     @SuppressWarnings("unchecked")
     @PostMapping(path = "/{roleId}/menus")
     public ResponseEntity<Role> addRoleMenus(@PathVariable String roleId, @RequestBody Map<String, Object> reqMap) {
-        // TODO 具体业务实现
+        userService.checkPermission("/auth/roles/" + roleId + "/menus", RequestMethod.POST);
         String ids = (String)reqMap.get("ids");
         return responseOfPost(roleService.addRoleMenus(roleId, ids));
     }
 
     @DeleteMapping("/{roleId}/menus")
     public ResponseEntity deleteRoleMenus(@PathVariable String roleId, @RequestParam String ids) {
-        // TODO 具体业务实现
+        userService.checkPermission("/auth/roles/" + roleId + "/menus", RequestMethod.DELETE);
         return responseOfDelete(roleService.deleteRoleMenus(roleId, ids) != null);
     }
 
@@ -114,7 +120,7 @@ public class RolesController extends BaseController {
 
     @GetMapping(path = "/parents")
     public ResponseEntity<Collection<? extends Role>> getMenuParents() {
-        // TODO 具体实现 获取父节点菜单列表
+        userService.checkPermission("/auth/roles/parents", RequestMethod.GET);
         return responseOfGet(roleService.getRoleParents());
     }
 }
